backend/bookmarkit/requirements.txt
beautifulsoup4==4.12.2
django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
drf-spectacular==0.26.4
gunicorn==21.2.0
pillow==10.0.0
python-dotenv==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
requests==2.31.0
